services/BotService/pyproject.toml
[project]
name = "botservice"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "aiogram>=3.0",
    "python-dotenv>=1.0",
    "httpx>=0.27.0",
    "aiokafka>=0.11.0",
]
